Fix Alexa for Core 2.5.0 and backwards compatible
Alexa uses Phillips Hue Emulation to control Tasmota.

In the Arduino Esp8266 Core v2.3.0 all works fine.
In the Arduino Esp8266 Core v2.4.0, 2.4.1 and 2.4.2, a modification in the core is needed in order to make phillips hue work.
In the Arduino Esp8266 Core v2.5.0 (actual Stage version), the webserver parsing was modified and this PR fix this modification.

The new parsing of the core works fine for a standard phillips hue messages, but for Alexa it splits the command in 2 args. The first arg is empty and the second contains the correct phillips hue commands.