fix empty prefix issue
When switchtopic is set to a different topic the prefix will be empty so its length will be zero. In the Shorten function the strncmp will give a zero result because the prefixlen is zero and will add '~' even though there is no prefix.
Adding prefixlen != 0 in the if condition fixes the issue. Also skip adding the prefix "~" on line 292 by checking if the prefix length is greater than zero.