add -DBEARSSL_SSL_BASIC
When using TLS, this flag avoids loading all the Elliptic Curves algorithm, and saves ~45KB of flash. Anyways, Elliptic Curves are far too slow for ESP8266 and the handshake would take more the 5 seconds. With this flag we limit to RSA algorithms which are faster.
Thx for the tipp @s-hadinger