fix pwm multi channel (Option68)
With Option68 enabled, the pwm channels are independent. The channel
setting code was indirectly using setRGB and setCW. Those methods
recalculate the channel values to adjust for brightness, thereby
modifying the channel values.

This change add raw channel set/get methods that do not modify the
channels values and uses those when pwm_multichannel is in effect.