Add a key to the TuyaReceived to help when writing boilerplate rules.
The key is in the form Cmnd<msgtype>DpId<ID>DpType<DpIdType> with a
value that is the hex encoding of the DpIdData. This allows Rules to be
written that target status updates from a specific DpId and of a specific
message type. For example, an Anko HEGSM40 fan reports the speed the fan
is operating at via STAT message on DpId 4 of type Enum. To publish the
speed to MQTT as it's reported from the MCU, I can use the following rule:

  on TuyaReceived#Cmnd7DpId3DpType4 do Publish stat/%topic%/SPEED %value% endon

This appears "on the wire" as:

  stat/tasmota/SPEED 01