Fixed last_seen calculation with zigbee devices.
The specific circumstances of this bug typically involve a restart, after which a zigbee
device pings the bridge before the Rtc.utc_time is set by NTP. This results in a large "last seen"
time calculation. This is cosmetic only. This patch ensures the "last seen" time is after
2020-01-01 0000 UTC when pulled from the Rtc.utc_time field, otherwise, last_seen updates are
ignored.